Bugfix and Changes to MeanShiftClustering --- Deepak
Made some changes to use the newer version of kdtree. Please update the folder extern/kdtree_andrea to get the newer version which is faster and has a KDTree class which takes care of freeing the kdtree memory without you worrying about it.

Also, fixed a small bug in Optimized mode of the mean-shift. Once the mode seeking procedure converges, before deciding to create a new cluster i check if an existing cluster if closeby (within certain bounds). If one exists, i merge it with the closest cluster. If not. i create a new cluster. This was not being done properly in the previous code. So, i fixed it. Hopefully, the results should only get better. 



git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9431 0bcb941b-74d1-0310-bb26-e6636707f451