BugFix --- Deepak --- blobDetection
Fixed a bug with the help of francois caused by locmax3d. locmax3d was ignoring the seed point near the image border due to which i was missing a lot of blobs near image border in my data. Now locmax3d has a flag 'ClearBorder' which should be set to false if you dont want it to ignore stuff near the image border.

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9600 0bcb941b-74d1-0310-bb26-e6636707f451