Extended Functionality -- Deepak -- thresholding algorithms
In thresholdLocalSeg.m and thresholdLocalCalculate.m, i added a way to pass in a function object for the choice_of_threshold argument. This will let you use your favourite thresholding algorithm apart from the ones built into these functions (otsu, rosin, fluorecesce image).

In thresholdOtsu.m, i return the thresholded foreground mask as an optional output argument. I do realize that this function might be widely used by lots of people and in a lot of our packages. But, this change should not affect any existing code. Please feel 
free to scream at me if you want it reverted back, but i felt this feature is useful as we mostly are interested in the mask more than the threshold value.


git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9806 0bcb941b-74d1-0310-bb26-e6636707f451