Deepak -- Performance Bug Fix -- MeanShiftClustering.m
Fixed a bug that was causing "out of memory" on large data sets with a lot of clusters. Many thanks to jeff who brought this to my notice or i wouldnt even have noticed it. 

He had a dataset of approx 800,000 points which seemed very sparse and he had a lot of clusters (>10,000) in his data. This was eventually making the pointToClusterMap (dimensions -- numPoints x numClusters) to occupy a humungous amount of memory. For now, i fixed it by making it a sparse matrix (since this matrix is indeed very sparse). A more efficient way to implement this would be using a list of linked lists or as a cell array.




git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@10450 0bcb941b-74d1-0310-bb26-e6636707f451