Add function to cleanup ROI paackage with integration test
Input of the functions are a package name and an optional list of processes
to keep. The function goes through all the ROIs of the movie, unlinks the
parent package and the processes. If some processes are to be kept,
independent packages are created in the ROIs and the parent process is
remapped.

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@11137 0bcb941b-74d1-0310-bb26-e6636707f451