Add ButtonDownFcn property to LineDisplay to allow interaction
This commit adds an extra property to the LineDisplay allowing to pass a
function handle so that a callback can be associated with the drawn object.
The default property is set to empty but can be initialized by Process
classes. Additionally the display properties are moved to a standalone method
so that calling Process.draw() on an existing object can update it.

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@11259 0bcb941b-74d1-0310-bb26-e6636707f451