Add logic in BioformatsReader to store and set series index
With the implementation of shared readers, the index of the series can be
modified by another existing reader. This commit implements an internal
property storing the index of the reader and wraps all calls to the
Bio-Formats reader using getReader() which also sets the series. Tests are
written that use or not the shared reader to check the setSeries() logic.

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@11371 0bcb941b-74d1-0310-bb26-e6636707f451