Handle Orphaned images as a dataset in the OMERO data selection interface
In edge cases (Michael Porter), the orphaned images can be huge and the user
does not want this loaded by default for performance reasons. This commit
fixes this issue by treating the Orphaned images as a virtual dataset and
appending it to the popup menu when no project is selected.

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@11476 0bcb941b-74d1-0310-bb26-e6636707f451