Fix movieInfo initialization in point source detection process
This commit should fix a bug that arises when no particle is detected in the
first frame of a movie and pstruct returns zero. The movieInfo structure array
should be initialized to contain the minimum amount of fields required by the
tracker and then expanded when the first non-empty pstruct is returned.