Fully delegate the dimension check to MovieData.sanityCheck()
With the migration of the data loading/dimension check to the reader classes,
there is no more logic in Channel (only delegation). There is no more reason
the dimension check should be performed at that level. With the channels
dimension mismatch been handled at the Reader level (see
TiffSeries.getDimensions()) all the logic is now moved to
MovieData.sanityCheck().