Simplify logic of loadMatFile() methods
This commit modifies the low-level loadMatFile() methods to
1- take a single argument (full path of the MAT file) as an input
2- return an unchecked MovieData/MovieList object.
The sanity check should now performed post-loading at the level of the caller
(typically MovieData.load()/MovieList.load()).

For the OMERO connection, this commit also simplifies the logic with the
setOmeroSession() calls being only integrated in the omero functionalities.
This commit improves future improvements in terms of permissions for such
objects.